% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_fetch_flights.R
\name{check_chrome_installation}
\alias{check_chrome_installation}
\title{Check if Chrome/Chromium is installed}
\usage{
check_chrome_installation(verbose = TRUE)
}
\description{
Check if Chrome/Chromium is installed
}
\keyword{internal}
