% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_fetch_flights.R
\name{check_internet_connection}
\alias{check_internet_connection}
\title{Check internet connection}
\usage{
check_internet_connection(verbose = TRUE)
}
\description{
Check internet connection
}
\keyword{internal}
