% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight.R
\name{classify_arg}
\alias{classify_arg}
\title{Classify Flight Argument}
\usage{
classify_arg(flight, arg)
}
\arguments{
\item{flight}{Flight object being constructed}

\item{arg}{Argument to classify}
}
\description{
Classify Flight Argument
}
\keyword{internal}
