% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{clean_results}
\alias{clean_results}
\title{Clean and parse results from scraped page}
\usage{
clean_results(result, date, verbose = TRUE)
}
\description{
Clean and parse results from scraped page
}
\keyword{internal}
