% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_fetch_flights.R
\name{close_chromote_safely}
\alias{close_chromote_safely}
\title{Close chromote browser safely}
\usage{
close_chromote_safely(browser)
}
\description{
Close chromote browser safely
}
\keyword{internal}
