% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight.R
\name{flights_to_dataframe}
\alias{flights_to_dataframe}
\title{Convert Flight objects to data frame}
\usage{
flights_to_dataframe(flights)
}
\arguments{
\item{flights}{List of Flight objects}
}
\value{
A data frame with flight information
}
\description{
Converts a list of Flight objects into a data frame.
}
\examples{
\dontrun{
flight1 <- Flight("2025-12-25", "JFKIST", "$450", "Nonstop")
flight2 <- Flight("2025-12-26", "ISTCDG", "$300", "1 stop")
flight3 <- Flight("2025-12-27", "CDGJFK", "$500", "Nonstop")
flights <- list(flight1, flight2, flight3)
df <- flights_to_dataframe(flights)
}
}
