% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{get_flight_elements_chromote}
\alias{get_flight_elements_chromote}
\title{Get flight elements from page using chromote}
\usage{
get_flight_elements_chromote(browser)
}
\description{
Get flight elements from page using chromote
}
\keyword{internal}
