% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_airport_conversion.R
\name{get_metropolitan_code}
\alias{get_metropolitan_code}
\title{Get Metropolitan Area Code for City Name}
\usage{
get_metropolitan_code(city_name)
}
\arguments{
\item{city_name}{Character string of a city name}
}
\value{
Metropolitan area code if it exists, NULL otherwise
}
\description{
Internal function to check if a city name has a corresponding metropolitan area code
(e.g., "New York" -> "NYC", "London" -> "LON"). These codes are used by airlines
and Google Flights to represent all airports in a metropolitan area.
}
\keyword{internal}
