% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_fetch_flights.R
\name{get_results_chromote}
\alias{get_results_chromote}
\title{Get results from a single URL using chromote}
\usage{
get_results_chromote(url, date, browser, verbose = TRUE)
}
\description{
Get results from a single URL using chromote
}
\keyword{internal}
