% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_define_query.R
\name{make_url}
\alias{make_url}
\title{Make URL for Google Flights Query}
\usage{
make_url(origins, dests, dates)
}
\arguments{
\item{origins}{List of origin airports}

\item{dests}{List of destination airports}

\item{dates}{List of dates}
}
\description{
Make URL for Google Flights Query
}
\keyword{internal}
