% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{make_url_request_chromote}
\alias{make_url_request_chromote}
\title{Make URL request and wait for content using chromote}
\usage{
make_url_request_chromote(url, browser, verbose = TRUE)
}
\description{
Make URL request and wait for content using chromote
}
\keyword{internal}
