% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_define_query.R
\name{normalize_query_args}
\alias{normalize_query_args}
\title{Normalize Query Arguments}
\usage{
normalize_query_args(args)
}
\arguments{
\item{args}{List of query arguments}
}
\value{
List of normalized arguments
}
\description{
Pre-processes query arguments to convert city names to airport codes.
Only processes string arguments that are not dates.
}
\keyword{internal}
