% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight.R
\name{parse_args}
\alias{parse_args}
\title{Parse Flight Arguments}
\usage{
parse_args(flight, args)
}
\arguments{
\item{flight}{Flight object being constructed}

\item{args}{List of arguments to parse}
}
\description{
Parse Flight Arguments
}
\keyword{internal}
