% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight.R
\name{print.Flight}
\alias{print.Flight}
\title{Print method for Flight objects}
\usage{
\method{print}{Flight}(x, ...)
}
\arguments{
\item{x}{A Flight object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for Flight objects
}
