% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{print.flight_query}
\alias{print.flight_query}
\title{Print method for flight query objects}
\usage{
\method{print}{flight_query}(x, ...)
}
\arguments{
\item{x}{A flight query object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for flight query objects
}
