% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight.R
\name{print.flight_record}
\alias{print.flight_record}
\title{Print method for flight_record objects}
\usage{
\method{print}{flight_record}(x, ...)
}
\arguments{
\item{x}{A flight_record object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for flight_record objects
}
