% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_fetch_flights.R
\name{print.flight_results}
\alias{print.flight_results}
\title{Print method for flight_results objects}
\usage{
\method{print}{flight_results}(x, ...)
}
\arguments{
\item{x}{A flight_results object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for flight_results objects
}
