% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{print.Scrape}
\alias{print.Scrape}
\title{Print method for Scrape objects}
\usage{
\method{print}{Scrape}(x, ...)
}
\arguments{
\item{x}{A Scrape object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for Scrape objects
}
