% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_multi_origin}
\alias{sample_multi_origin}
\title{Sample Multiple Origin Queries}
\format{
A named list of 2 flight_query objects (BOM and DEL to JFK)
\describe{
  \item{BOM}{query object for Mumbai (BOM) to JFK}
  \item{DEL}{query object for Delhi (DEL) to JFK}
}
}
\usage{
sample_multi_origin
}
\description{
Sample query objects for multiple origins created with `fa_define_query_range()`.
Demonstrates searching multiple airports over a date range.
}
\examples{
data(sample_multi_origin)
names(sample_multi_origin)
print(sample_multi_origin$BOM)
}
\keyword{datasets}
