% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_query}
\alias{sample_query}
\title{Sample Flight Query}
\format{
A flight_query object for JFK to IST round-trip
\describe{
  \item{origin}{List of origin airport codes}
  \item{dest}{List of destination airport codes}
  \item{dates}{List of travel dates}
  \item{type}{Trip type (round-trip)}
  \item{url}{Google Flights URLs}
}
}
\usage{
sample_query
}
\description{
A sample flight query object created with `fa_define_query()`.
Useful for testing and documentation examples without making API calls.
}
\examples{
data(sample_query)
str(sample_query)
print(sample_query)
}
\keyword{datasets}
