% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_data_chromote}
\alias{scrape_data_chromote}
\title{Scrape data for a single Scrape object using chromote}
\usage{
scrape_data_chromote(obj, browser, verbose = TRUE)
}
\description{
Scrape data for a single Scrape object using chromote
}
\keyword{internal}
