% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{set_properties}
\alias{set_properties}
\title{Set Scrape Properties}
\usage{
set_properties(scrape, args)
}
\arguments{
\item{scrape}{Scrape object being constructed}

\item{args}{List of arguments}
}
\description{
Set Scrape Properties
}
\keyword{internal}
