% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{initialize_chromote_browser}
\alias{initialize_chromote_browser}
\title{Initialize chromote browser}
\usage{
initialize_chromote_browser(headless = TRUE)
}
\description{
Initialize chromote browser
}
\keyword{internal}
