% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{set_properties}
\alias{set_properties}
\title{Set Query Properties}
\usage{
set_properties(query, args)
}
\arguments{
\item{query}{Query object being constructed}

\item{args}{List of arguments}
}
\description{
Set Query Properties
}
\keyword{internal}
